% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_design_patterns.R
\name{mod_design_patterns_ui}
\alias{mod_design_patterns_ui}
\title{Design Patterns Module}
\usage{
mod_design_patterns_ui(id)
}
\description{
A module to display the design patterns and technical implementation details
of the application.
}
